Moved MainActivity content into MainFragment
