+++ example/android/app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:orientation="vertical" android:layout_width="match_parent"
+              android:layout_height="match_parent">
+    <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="58dp"
+            android:titleTextColor="#ffffff"/>
+    <Button
+            android:id="@+id/button_screen"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+    <Button
+            android:id="@+id/button_tab"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+</LinearLayout>
